feat: data viz by geo - updated api call to include state
